moved init under apps, added a --force option for future-proofing, package.json now has a placeholder for the start script
