+++ sectionctl.go
@@ -29,7 +29,6 @@ type CLI struct {
 	Ps                 commands.PsCmd               `cmd help:"Show status of running applications"`
 	Version            commands.VersionCmd          `cmd help:"Print sectionctl version"`
 	WhoAmI             commands.WhoAmICmd           `cmd name:"whoami" help:"Show information about the currently authenticated user"`
-	Init               commands.InitCmd             `cmd help:"Initilize your project for deployment"`
 	Debug              bool                         `env:"DEBUG" help:"Enable debug output"`
 	SectionToken       string                       `env:"SECTION_TOKEN" help:"Secret token for API auth"`
 	SectionAPIPrefix   *url.URL                     `default:"https://aperture.section.io" env:"SECTION_API_PREFIX"`
